perf: add load test for active orders
